chore: rename api token service file to use camelcase
