+++ src/lib/services/api-token-service.limit.test.ts
@@ -2,7 +2,7 @@ import { createTestConfig } from '../../test/config/test-config';
 import type { IUnleashConfig } from '../server-impl';
 import { ApiTokenType } from '../types/models/api-token';
 import { ExceedsLimitError } from '../error/exceeds-limit-error';
-import { createFakeApiTokenService } from '../features/api-tokens/create-api-token-service';
+import { createFakeApiTokenService } from '../features/api-tokens/createApiTokenService';
 
 const createServiceWithLimit = (limit: number) => {
     const config: IUnleashConfig = createTestConfig({
